+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Perf
 
+### 2025-08-28
+
+- Improve precompiles further: modexp, ecrecover [#4168](https://github.com/lambdaclass/ethrex/pull/4168)
+
 ### 2025-08-27
 
 - Improve memory resize performance [#4117](https://github.com/lambdaclass/ethrex/pull/4177)